Move return_status to models
